feat: paginate display cells
