+++ config/routes.rb
@@ -25,6 +25,8 @@ Rails.application.routes.draw do
       resources :block_transactions, only: :show
       resources :addresses, only: :show
       get "/transactions/:id", to: "ckb_transactions#show", as: "ckb_transaction"
+      get "/transactions/:id/display_inputs", to: "ckb_transactions#display_inputs"
+      get "/transactions/:id/display_outputs", to: "ckb_transactions#display_outputs"
       get "/transactions", to: "ckb_transactions#index", as: "ckb_transactions"
       post "/transactions/query", to: "ckb_transactions#query", as: "query_ckb_transactions"
       resources :cell_input_lock_scripts, only: :show
